move business code to server/lib/
