+++ server/__tests__/config-test.js
@@ -13,9 +13,9 @@ const ship = {
 describe("buildConfigFromShip", () => {
   it("build a config object from the ship config object", () => {
 
-    const config = require("../config");
-    const getServiceAttributeToHullTopLevel = require("../mapping-data").getServiceAttributeToHullTopLevel;
-    const getServiceAttributeToHullTrait = require("../mapping-data").getServiceAttributeToHullTrait;
+    const config = require("../lib/config");
+    const getServiceAttributeToHullTopLevel = require("../lib/mapping-data").getServiceAttributeToHullTopLevel;
+    const getServiceAttributeToHullTrait = require("../lib/mapping-data").getServiceAttributeToHullTrait;
 
     const c = config.buildConfigFromShip(ship, organization, secret);
     expect(_.keys(c)).toEqual(["hull", "settings", "salesforce", "sync", "mappings"]);
